Use new upsert subscription logic
