+++ services/external/stripe/actions/upsert-subscription.ts
@@ -1,35 +1,75 @@
-import {
-	db,
-	stripeUserMappings,
-	subscriptions,
-	teamMemberships,
-	teams,
-	users,
-} from "@/drizzle";
+import { db, draftTeams, subscriptions, teams } from "@/drizzle";
+import { DRAFT_TEAM_METADATA_KEY } from "@/services/teams/actions/create-team";
 import { eq } from "drizzle-orm";
+import type Stripe from "stripe";
 import { stripe } from "../config";
 
 const timestampToDateTime = (timestamp: number) => new Date(timestamp * 1000);
 
-export const upsertSubscription = async (
-	subscriptionId: string,
-	customerId: string,
-) => {
+export const upsertSubscription = async (subscriptionId: string) => {
 	const subscription = await stripe.subscriptions.retrieve(subscriptionId);
+	const existingSubscriptionRecord = await db
+		.select()
+		.from(subscriptions)
+		.where(eq(subscriptions.id, subscription.id));
 
-	// TODO: When team plans are released, a user may belong to multiple teams, so we need to handle that case.
-	const [team] = await db
-		.select({ dbId: teams.dbId })
-		.from(teams)
-		.innerJoin(teams, eq(teams.dbId, subscriptions.teamDbId))
-		.innerJoin(teamMemberships, eq(teamMemberships.teamDbId, teams.dbId))
-		.innerJoin(users, eq(users.dbId, teamMemberships.userDbId))
-		.innerJoin(stripeUserMappings, eq(stripeUserMappings.userDbId, users.dbId))
-		.where(eq(stripeUserMappings.stripeCustomerId, customerId));
+	if (existingSubscriptionRecord.length > 0) {
+		await updateSubscription(subscription);
+		return;
+	}
 
-	const upsertValues: typeof subscriptions.$inferInsert = {
+	await activateProTeamSubscription(subscription);
+};
+
+async function activateProTeamSubscription(subscription: Stripe.Subscription) {
+	if (!(DRAFT_TEAM_METADATA_KEY in subscription.metadata)) {
+		throw new Error("Draft team ID not found in subscription metadata");
+	}
+	const draftTeamDbId = Number.parseInt(
+		subscription.metadata[DRAFT_TEAM_METADATA_KEY],
+		10,
+	);
+	const teamDbId = await createTeamFromDraftTeam(draftTeamDbId);
+	await insertSubscription(subscription, teamDbId);
+}
+
+async function createTeamFromDraftTeam(draftTeamDbId: number) {
+	const teamDbId = await db.transaction(async (tx) => {
+		const records = await db
+			.select()
+			.from(draftTeams)
+			.for("update")
+			.where(eq(draftTeams.dbId, draftTeamDbId));
+		if (records.length === 0) {
+			throw new Error(`Draft team with ID ${draftTeamDbId} not found`);
+		}
+
+		const [draftTeam] = records;
+		if (draftTeam.teamDbId) {
+			// team has already been created
+			return draftTeam.teamDbId;
+		}
+
+		const [team] = await db
+			.insert(teams)
+			.values({
+				name: draftTeam.name,
+				isInternalTeam: false,
+			})
+			.returning({ dbid: teams.dbId });
+		return team.dbid;
+	});
+
+	return teamDbId;
+}
+
+async function insertSubscription(
+	subscription: Stripe.Subscription,
+	teamDbId: number,
+) {
+	await db.insert(subscriptions).values({
 		id: subscription.id,
-		teamDbId: team.dbId,
+		teamDbId: teamDbId,
 		status: subscription.status,
 		cancelAtPeriodEnd: subscription.cancel_at_period_end,
 		cancelAt:
@@ -55,9 +95,40 @@ export const upsertSubscription = async (
 			subscription.trial_end !== null
 				? timestampToDateTime(subscription.trial_end)
 				: null,
-	};
-	await db.insert(subscriptions).values(upsertValues).onConflictDoUpdate({
-		target: subscriptions.id,
-		set: upsertValues,
 	});
-};
+}
+
+async function updateSubscription(subscription: Stripe.Subscription) {
+	await db
+		.update(subscriptions)
+		.set({
+			status: subscription.status,
+			cancelAtPeriodEnd: subscription.cancel_at_period_end,
+			cancelAt:
+				subscription.cancel_at !== null
+					? timestampToDateTime(subscription.cancel_at)
+					: null,
+			canceledAt:
+				subscription.canceled_at !== null
+					? timestampToDateTime(subscription.canceled_at)
+					: null,
+			currentPeriodStart: timestampToDateTime(
+				subscription.current_period_start,
+			),
+			currentPeriodEnd: timestampToDateTime(subscription.current_period_end),
+			created: timestampToDateTime(subscription.created),
+			endedAt:
+				subscription.ended_at !== null
+					? timestampToDateTime(subscription.ended_at)
+					: null,
+			trialStart:
+				subscription.trial_start !== null
+					? timestampToDateTime(subscription.trial_start)
+					: null,
+			trialEnd:
+				subscription.trial_end !== null
+					? timestampToDateTime(subscription.trial_end)
+					: null,
+		})
+		.where(eq(subscriptions.id, subscription.id));
+}
